requirements.txt
sentence-transformers
faiss-cpu
ollama